QRN-2 - fix empty file issue in TheHerbWitches - make some separation of logic - add dummy Ingredient component
