+++ src/components/quacksOfQuedlingburgRandomizer/UI/Ingredient/Ingredient.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Ingredient = () => {
+    return(
+            <div>wee</div>
+    )
+};
+
+export default Ingredient;
