+++ src/components/quacksOfQuedlingburgRandomizer/constants/gameEntities.ts
@@ -1,0 +1,27 @@
+const INGREDIENT_BOOK_NAMES = [
+    'Crow Skull',
+    'Toadstool',
+    'Mandrake',
+    'Garden Spider',
+    'Ghost\'s Breath',
+    'Locoweed',
+] as const;
+
+const SPECIAL_INGREDIENT_BOOK_NAMES = [
+    'Pumpkin',
+    'African Death\'s Head Hawkmoth',
+];
+
+const BOOK_SET_VALUES = [1, 2, 3, 4, 5, 6];
+
+const EXTENSIONS = [
+    'The Herb Witches',
+    'The Alchemists',
+];
+
+export {
+    INGREDIENT_BOOK_NAMES,
+    SPECIAL_INGREDIENT_BOOK_NAMES,
+    BOOK_SET_VALUES,
+    EXTENSIONS,
+};
